+++ src/SDLRenderWrapper.h
@@ -15,7 +15,7 @@ public:
 	void ClearFromTextureMap(std::string filename) override;
 	void ClearAllTextures() override;
 	void RenderText(const std::string& text, const std::string& fontFile, int fontSize, int x, int y, double scale = 1, bool centered = 0, std::optional<Color> color = std::nullopt) override;
-	void DrawRectangle(int x, int y, int width, int height, Color color) override;
+	void DrawRectangle(int x, int y, float width, float height, Color color) override;
 private:
 	SDL_Renderer* renderer;
 	SDL_Window* window;
